model: Move AnalyzerConfiguration to config package
The class should be in the same package as the other configuration classes.
