+++ analyzer/src/funTest/kotlin/GoDepTest.kt
@@ -21,10 +21,10 @@ package com.here.ort.analyzer
 
 import com.here.ort.analyzer.managers.GoDep
 import com.here.ort.downloader.VersionControlSystem
-import com.here.ort.model.AnalyzerConfiguration
 import com.here.ort.model.Identifier
 import com.here.ort.model.Project
 import com.here.ort.model.VcsInfo
+import com.here.ort.model.config.AnalyzerConfiguration
 import com.here.ort.model.yamlMapper
 import com.here.ort.utils.test.USER_DIR
