+++ analyzer/src/funTest/kotlin/SbtTest.kt
@@ -21,7 +21,7 @@ package com.here.ort.analyzer
 
 import com.here.ort.analyzer.managers.SBT
 import com.here.ort.downloader.vcs.Git
-import com.here.ort.model.AnalyzerConfiguration
+import com.here.ort.model.config.AnalyzerConfiguration
 import com.here.ort.model.yamlMapper
 import com.here.ort.utils.test.patchExpectedResult
