+++ analyzer/src/main/kotlin/Main.kt
@@ -26,8 +26,8 @@ import com.beust.jcommander.JCommander
 import com.beust.jcommander.Parameter
 import com.beust.jcommander.ParameterException
 
-import com.here.ort.model.AnalyzerConfiguration
 import com.here.ort.model.OutputFormat
+import com.here.ort.model.config.AnalyzerConfiguration
 import com.here.ort.utils.PARAMETER_ORDER_HELP
 import com.here.ort.utils.PARAMETER_ORDER_LOGGING
 import com.here.ort.utils.PARAMETER_ORDER_MANDATORY
