+++ analyzer/src/main/kotlin/managers/Bower.kt
@@ -21,7 +21,7 @@ package com.here.ort.analyzer.managers
 
 import com.here.ort.analyzer.PackageManager
 import com.here.ort.analyzer.PackageManagerFactory
-import com.here.ort.model.AnalyzerConfiguration
+import com.here.ort.model.config.AnalyzerConfiguration
 
 import java.io.File
