+++ model/src/main/kotlin/Mappers.kt
@@ -29,6 +29,9 @@ import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
 
+import com.here.ort.model.config.AnalyzerConfiguration
+import com.here.ort.model.config.AnalyzerConfigurationDeserializer
+
 val ortModelModule = SimpleModule("OrtModelModule").apply {
     addDeserializer(AnalyzerConfiguration::class.java, AnalyzerConfigurationDeserializer())
     addDeserializer(Error::class.java, ErrorDeserializer())
